replaced import parser with a module
