+++ spec/support/string_tool.rb
